+++ final-tg-bot/src/main/java/org/example/finaltgbot/dto/telegram/Result.java
@@ -6,6 +6,6 @@ import lombok.Setter;
 @Getter
 @Setter
 public class Result {
-    public int update_id;
+    public Long update_id;
     public Message message;
 }
